Remove link to email in Contact section.
This change is in response to comment about an unused CSS id. This link was taken out in a latter branch commit as the sidebar already provides access to my email to contact me. A latter addition will include an arrow to the sidebar or pull the sidebar icons to the bottom regardless of the window side when the user scrolls to the 'Contact' page.