Add comment add and comment thread functionality.
Created a barebones comments system that allows a user to enter a comment in a text form. Once submit is clicked, a POST request is sent to the server which saves the string in the comments to a List of strings. After being redirected back to the original page, a get request is sent to retreive all of the comments stored on the server. These are fetched and served to the UI with JavaScript.