Add functionality to load comments from Datastore.
The one java servlet was split into two servlets: one responsible for post requests (NewComment) and one for get requests (ListComments). NewComment enters comments that were submitted via the html form into the datastore as a new entity of the Comment kind. Once re-directed to the page, the ListComment servlet is called via a get request. A query is made to the datastore for all of the entities in the Comment kind. These comments are converted to a list contained in a JSON string and written to a page that the javascript fetches from.