Store and serve images using their blob key vs a generated URL.
Previously ImageService.getServingUrl was used with the blob key for an image to generate a URL for the uploaded image. This was stored in the database and fetched from JS where is was inserted as the value of the src tag in the comment. Unfortunately due to access issues on the cloud project, this did not work when deploying the server. So rather than storing and serving the image URL, the blob key was stored and served 'manually.' Once the blobkey was stored and fetched on the JS side, the blobkey was added as a parameter to a query string which fetched the file from blobstore via a new servlet.